==> notebooks/14_inference_pipeline.ipynb <==
import sys
import os
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))

%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow()).floor('h') # - timedelta(hours=1)
print(f'{current_date=}')

# # from src.feature_store_api import load_batch_of_features_from_store

# features = load_batch_of_features_from_store(current_date,n_features=5)

from src.feature_store_api import load_batch_of_features_from_store
from src import config
from src.config import FEATURE_VIEW_METADATA
features = load_batch_of_features_from_store(
    feature_view_metadata=config.FEATURE_VIEW_METADATA,
    n_features=5
)

print(features.head())

import hopsworks
from src import config

# Login to Hopsworks
project = hopsworks.login(project=config.HOPSWORKS_PROJECT_NAME)

# Project info
print("Connected to Hopsworks project:", project.name)

# Access model registry
mr = project.get_model_registry()

# If you know your model name, fetch it
model_name = "taxi_demand_predictor_next_hour"  # or whatever is in your project
try:
    models = mr.get_models(model_name)
    print(f"Models with name '{model_name}':", models)
except Exception as e:
    print(f"Could not fetch models: {e}")


from src.inference import load_model, get_model_predictions
from src.feature_store_api import load_batch_of_features_from_store
from src import config

# Load features
features = load_batch_of_features_from_store(config.FEATURE_VIEW_METADATA, config.N_FEATURES)

# Load model
model = load_model()

# Generate predictions
predictions = get_model_predictions(model, features)

features['predicted_demand'] = predictions['predicted_demand']
print(features.head())

predictions['pickup_hour'] = current_date
predictions

from src.feature_store_api import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Predictions generate by our production model",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)
